RF | Combining UserTest classes
